WIP: integrate new request builder to http client
