+++ client.go
@@ -3,7 +3,6 @@ package http
 
 import (
 	"bufio"
-	"bytes"
 	"context"
 	"fmt"
 	"io"
@@ -24,9 +23,6 @@ import (
 	"go.unistack.org/micro/v4/selector"
 	"go.unistack.org/micro/v4/semconv"
 	"go.unistack.org/micro/v4/tracer"
-	"google.golang.org/protobuf/proto"
-
-	"go.unistack.org/micro-client-http/v4/builder"
 )
 
 var DefaultContentType = "application/json"
@@ -571,189 +567,4 @@ func (c *Client) parseRsp(ctx context.Context, hrsp *http.Response, rsp any, opt
 
 	// if the error map item does not implement the error interface, wrap it
 	return &Error{err: mappedErr}
-}
-
-func buildHTTPRequest(
-	ctx context.Context,
-	addr string,
-	path string,
-	ct string,
-	cf codec.Codec,
-	msg any,
-	opts client.CallOptions,
-	log logger.Logger,
-) (
-	*http.Request,
-	error,
-) {
-	protoMsg, ok := msg.(proto.Message)
-	if !ok {
-		return nil, errors.BadRequest("go.micro.client", "msg must be a proto message type")
-	}
-
-	var (
-		method  = http.MethodPost
-		bodyOpt = "*"
-
-		parameters = map[string]map[string]string{}
-	)
-
-	if opts.Context != nil {
-		if v, ok := methodFromOpts(opts); ok {
-			method = v
-		}
-		if v, ok := pathFromOpts(opts); ok {
-			path = v
-		}
-		if v, ok := bodyFromOpts(opts); ok {
-			bodyOpt = v
-		}
-		if h, ok := headerFromOpts(opts); ok && len(h) > 0 {
-			m, ok := parameters["header"]
-			if !ok {
-				m = make(map[string]string)
-				parameters["header"] = m
-			}
-			for idx := 0; idx+1 < len(h); idx += 2 {
-				m[h[idx]] = h[idx+1]
-			}
-		}
-		if c, ok := cookieFromOpts(opts); ok && len(c) > 0 {
-			m, ok := parameters["cookie"]
-			if !ok {
-				m = make(map[string]string)
-				parameters["cookie"] = m
-			}
-			for idx := 0; idx+1 < len(c); idx += 2 {
-				m[c[idx]] = c[idx+1]
-			}
-		}
-	}
-
-	reqBuilder, err := builder.NewRequestBuilder(path, method, bodyOpt, protoMsg)
-	if err != nil {
-		return nil, errors.BadRequest("go.micro.client", "new request builder: %+v", err)
-	}
-
-	resolvedPath, newMsg, err := reqBuilder.Build()
-	if err != nil {
-		return nil, errors.BadRequest("go.micro.client", "request build: %+v", err)
-	}
-
-	u, err := url.Parse(fmt.Sprintf("%s%s", addr, resolvedPath))
-	if err != nil {
-		return nil, errors.BadRequest("go.micro.client", "%+v", err)
-	}
-
-	reqBody, err := cf.Marshal(newMsg)
-	if err != nil {
-		return nil, errors.BadRequest("go.micro.client", "%+v", err)
-	}
-
-	var hreq *http.Request
-
-	if len(reqBody) > 0 {
-		hreq, err = http.NewRequestWithContext(ctx, method, u.String(), io.NopCloser(bytes.NewBuffer(reqBody)))
-		hreq.ContentLength = int64(len(reqBody))
-	} else {
-		hreq, err = http.NewRequestWithContext(ctx, method, u.String(), nil)
-	}
-
-	if err != nil {
-		return nil, errors.BadRequest("go.micro.client", "%+v", err)
-	}
-
-	setHeadersAndCookies(ctx, hreq, ct, opts)
-	if err = validateHeadersAndCookies(hreq, parameters); err != nil {
-		return nil, errors.BadRequest("go.micro.client", "%+v", err)
-	}
-
-	if log.V(logger.DebugLevel) {
-		log.Debug(
-			ctx,
-			fmt.Sprintf("request %s to %s with headers %v body %s", method, u.String(), hreq.Header, reqBody),
-		)
-	}
-
-	return hreq, nil
-}
-
-func setHeadersAndCookies(ctx context.Context, r *http.Request, ct string, opts client.CallOptions) {
-	r.Header = make(http.Header)
-
-	r.Header.Set(metadata.HeaderContentType, ct)
-	r.Header.Set("Content-Length", fmt.Sprintf("%d", r.ContentLength))
-
-	if opts.AuthToken != "" {
-		r.Header.Set(metadata.HeaderAuthorization, opts.AuthToken)
-	}
-
-	if opts.StreamTimeout > time.Duration(0) {
-		r.Header.Set(metadata.HeaderTimeout, fmt.Sprintf("%d", opts.StreamTimeout))
-	}
-	if opts.RequestTimeout > time.Duration(0) {
-		r.Header.Set(metadata.HeaderTimeout, fmt.Sprintf("%d", opts.RequestTimeout))
-	}
-
-	if opts.RequestMetadata != nil {
-		for k, v := range opts.RequestMetadata {
-			if k == "Cookie" {
-				applyCookies(r, v)
-				continue
-			}
-			r.Header[k] = append(r.Header[k], v...)
-		}
-	}
-
-	if md, ok := metadata.FromOutgoingContext(ctx); ok {
-		for k, v := range md {
-			if k == "Cookie" {
-				applyCookies(r, v)
-				continue
-			}
-			r.Header[k] = append(r.Header[k], v...)
-		}
-	}
-}
-
-func applyCookies(r *http.Request, rawCookies []string) {
-	if len(rawCookies) == 0 {
-		return
-	}
-
-	raw := strings.Join(rawCookies, "; ")
-
-	tmp := http.Request{Header: http.Header{}}
-	tmp.Header.Set("Cookie", raw)
-
-	for _, c := range tmp.Cookies() {
-		r.AddCookie(c)
-	}
-}
-
-func validateHeadersAndCookies(r *http.Request, parameters map[string]map[string]string) error {
-	if headers, ok := parameters["header"]; ok {
-		for name, required := range headers {
-			if required == "true" && r.Header.Get(name) == "" {
-				return fmt.Errorf("missing required header: %s", name)
-			}
-		}
-	}
-
-	if cookies, ok := parameters["cookie"]; ok {
-		cookieMap := map[string]string{}
-		for _, c := range r.Cookies() {
-			cookieMap[c.Name] = c.Value
-		}
-
-		for name, required := range cookies {
-			if required == "true" {
-				if _, ok := cookieMap[name]; !ok {
-					return fmt.Errorf("missing required cookie: %s", name)
-				}
-			}
-		}
-	}
-
-	return nil
 }
